Use a trait to build Calendar Events
Both User and School events should contain the same data and be built
the same way.
